Move CombineOptions class into iris.loading to avoid circularity.
